story-to-video/requirements.txt
click>=8.1.0
pydantic>=2.0.0
openai>=1.0.0
pyyaml>=6.0
requests>=2.31.0
